\name{crashDat}
\docType{data}
\alias{crashDat}
\title{Crash Dataset: Calibration of Crash Dummies in Automobile Safety Tests}
\description{Data from 58 simulated car crashes were analyzed. The relationship between the crash outcome (fatal, non-fatal) and 3 covariates was modeled.}
\usage{data(drugDat)}
\format{Matrix with columns:
\tabular{rlll}{
[,1] \tab age \tab numeric \tab designed age of the crash dummy  \cr
[,2] \tab vel \tab numeric \tab velocity of the car at impact \cr
[,3] \tab acl \tab numeric \tab acceleration of the car at impact \cr
[,4] \tab y \tab numeric \tab outcome from test (1=fatal, 0=non-fatal) \cr
[,5] \tab n \tab numeric \tab number of binomial trials. For this data, n is a vector of ones, thus each trial is modeled as a Bernoulli random variable. \cr}
}
\keyword{datasets}
