\name{diabDat}
\docType{data}
\alias{diabDat}
\title{Simulated Diabetes Dataset}
\description{This dataset was simulated by sampling cases from an existing type 1 diabetes study (the original data could not be disclosed). The study investigates the relationship between concentration levels (low and high) of the islet antigen 2 antibody (IA2A) and several covariates of potential interest in type 1 diabetes patients (age, gender, and the number of copies (0,1 or 2) of the HLA-DQ2, HLA-DQ8 and HLA-DQ6.2 haplotypes).}
\usage{data(diabDat)}
\format{Matrix with columns:
\tabular{rlll}{
[,1] \tab n \tab numeric \tab number of binomial trials. \cr
[,2] \tab IA2A \tab numeric \tab number of patients with high concentration levels of the islet antigen 2 antibody \cr
[,3] \tab gender \tab numeric \tab gender of patient \cr
[,4] \tab age \tab numeric \tab age of patient  \cr
[,5] \tab nDQ2 \tab numeric \tab number of copies of the HLA-DQ2 haplotype (0,1, or 2) \cr
[,6] \tab nDQ8 \tab numeric \tab number of copies of the HLA-DQ8 haplotype (0,1, or 2) \cr
[,7] \tab nDQ6.2 \tab numeric \tab number of copies of the HLA-DQ6.2 haplotype (0,1, or 2) \cr}
}
\keyword{datasets}
