\name{drugDat}
\docType{data}
\alias{drugDat}
\title{Drug Dataset}
\description{Simulated data for a hypothetical drug experiment comparing control versus treatment.}
\usage{data(drugDat)}
\format{Matrix with columns:
\tabular{rlll}{
[,1] \tab sex \tab numeric \tab gender (1=male, 0=female) \cr
[,2] \tab treatment \tab numeric \tab treatment type (1=treatment, 0=control) \cr
[,3] \tab recovered \tab numeric \tab number of subjects that recovered \cr
[,4] \tab n \tab numeric \tab number of binomial trials }
}
\keyword{datasets}
