\name{titanDat}
\docType{data}
\alias{titanDat}
\title{Titanic Dataset}
\description{Relationship between survival and passenger class on the Titanic. The records of the sinking of the Titanic were studied to establish the relationship between survival and passenger class on the ship. For each person on board the ocean liner, this dataset records sex, age (child/adult), class (crew, 1st, 2nd, 3rd class) and whether or not the person survived.}
\usage{data(titanDat)}
\format{Matrix with columns:
\tabular{rlll}{
[,1] \tab surv \tab numeric \tab number of survivors \cr
[,2] \tab n \tab numeric \tab  total number of people \cr
[,3] \tab class \tab numeric \tab passenger class (0 = crew, 1 = first, 2 = second, 3 = third) \cr
[,4] \tab age \tab numeric \tab age group (1 = adult, 0 = child)) \cr
[,5] \tab sex \tab numeric \tab gender (1 = male, 0 = female) \cr }
}
\references{Mehta CR, Patel NR. Logxact for Windows. Cytel Software Corporation, Cambridge, USA, 1999.} 
\keyword{datasets}
