\name{utiDat}
\docType{data}
\alias{utiDat}
\title{Urinary Tract Infection and Contraceptive Use}
\description{How is the development of first-time urinary tract infection (UTI) related to contraceptive use? A study of sexually active college women with UTI was conducted, and their use of various contraceptives was surveyed.}
\usage{data(utiDat)}
\format{Matrix with columns:
\tabular{rlll}{
[,1] \tab uti \tab numeric \tab infection status (1=yes, 0=no) \cr
[,2] \tab n \tab numeric \tab number of binomial trials \cr
[,3] \tab age \tab numeric \tab age of the person \cr
[,4] \tab current \tab numeric \tab no regular current sex partner (1=yes,0=no) \cr
[,5] \tab dia \tab numeric \tab diaphragm use (1=yes, 0=no) \cr
[,6] \tab oc \tab numeric \tab oral contraceptive (1=yes, 0=no) \cr
[,7] \tab pastyr \tab numeric \tab no regular partner with relationship < 1yr  (1=yes, 0=no) \cr
[,8] \tab vi \tab numeric \tab vaginal intercourse (1=yes, 0=no) \cr
[,9] \tab vic \tab numeric \tab vaginal intercourse with condom  (1=yes, 0=no) \cr
[,10] \tab vicl \tab numeric \tab vaginal intercourse with lubricated condom  (1=yes, 0=no) \cr
[,11] \tab vis \tab numeric \tab vaginal intercourse with spermicide (1=yes, 0=no) \cr}
}
\references{Mehta CR, Patel NR. Logxact for Windows. Cytel Software Corporation, Cambridge, USA, 1999.} 
\keyword{datasets}
